move routes out to route file
